First big commit, read me
Ok, so we are finally aproaching another release date.
The next release, probably is going to be 2.0 because of soo many
structural changes in the code architeture (and there's going to be
another minor architeture change, sorry).
However, these big changes are for good and the app is going to be
really awesome!