Added retrying pipeline delay on the config files
In the front-camera-MM-blobs-case, we create almost a low-level camera pipeline (!!). So, when taking long-exposure pics, we need to actually wait for the exposure time to pass (beforing calling Camera.release) to get a proper exposure. Since we can't (reliably in every-corner-case-possible) know the exposure time before taking the picture (in the weird i3av4 format), we always wait for the maximum long exposure possible. That's called the retrying pipeline delay.
Note: this is EXTREMELY HACKY!