Created CoarseIntegrationTimeMeteringService
As analog/digital gain "is" ISO, coarse/fine integration time "is" exposure time.
The gain->iso conversion needs a "base iso" value that depends of the sensor.
In a similar way, the integration time->exposure time also needs a time-scale value that depends of the sensor.
This new service enables full exposure metering on the front camera, closing the feature gap between the front/rear cameras